+++ src/cli.ts
@@ -1,3 +1,4 @@
+#!/usr/bin/env node
 import { runCliMain } from "./cli-main.js";
 
 void runCliMain({